bot/cryptocom-ai-telegram-bot/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
bitarray==3.4.2
black==24.10.0
Bottleneck==1.5.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
ckzg==2.1.1
click==8.2.1
crypto-com-developer-platform-client==1.1.0
cryptocom-agent-client==1.1.1
cytoolz==1.0.1
defusedxml==0.7.1
discord.py==2.5.2
distro==1.9.0
docstring_parser==0.16
eth-account==0.13.6
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
filelock==3.18.0
filetype==1.2.0
fireworks-ai==0.15.12
flake8==7.3.0
frozenlist==1.7.0
fsspec==2025.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.93.0
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.21.1
google-generativeai==0.8.4
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
groq==0.24.0
grpc-google-iam-v1==0.14.2
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.16.0
hexbytes==1.3.1
hf-xet==1.1.5
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.2
huggingface-hub==0.33.0
idna==3.10
iniconfig==2.1.0
isort==5.13.2
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-anthropic==0.3.1
langchain-core==0.3.55
langchain-fireworks==0.2.6
langchain-google-genai==2.0.8
langchain-google-vertexai==2.0.24
langchain-groq==0.3.2
langchain-mistralai==0.2.4
langchain-openai==0.3.12
langchain-text-splitters==0.3.8
langchain-xai==0.2.3
langfuse==2.57.5
langgraph==0.2.61
langgraph-checkpoint==2.1.0
langgraph-sdk==0.1.70
langsmith==0.3.45
mccabe==0.7.0
multidict==6.5.0
mypy_extensions==1.1.0
numexpr==2.11.0
numpy==2.2.6
openai==1.74.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
parsimonious==0.10.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycryptodome==3.23.0
pydantic==2.10.4
pydantic_core==2.27.2
pyflakes==3.4.0
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.11.1
pytz==2025.2
pyunormalize==16.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rlp==4.1.0
rsa==4.9.1
setuptools==78.1.1
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
tenacity==8.5.0
tiktoken==0.9.0
tokenizers==0.21.1
toolz==1.0.0
tqdm==4.67.1
types-requests==2.32.4.20250611
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.5.0
validators==0.35.0
web3==7.10.0
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.1
zstandard==0.23.0
